Remove create_before_destroy from network resources
Origin of this lifecycle configuration is a bit hazy, and we most
likely don't need it currently, and if it would turn useful, can add
it back anyway :)

The create_before_destroy[1] instructs Terraform to create a new
resource before destroying the old one, when the old resource cannot
be modified in-place. Not all resources of same type, with same
attribute values, can co-exist at the same time, because attributes
might be unique for example, so this behaviour is not used by default.

Also, removal of VPC resources is usually tricky, since usually
everything in the infrastructure depends on the VPC. So removal would
require deleting all other depending resources.

[1] https://www.terraform.io/docs/language/meta-arguments/lifecycle.html#create_before_destroy